n.st-0.1133djl

+++ src/main/java/com/crypto/analysis/main/core/data_utils/normalizers/MaxAbsScaler.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 import java.util.HashMap;
 
 public class MaxAbsScaler {
-    private final HashMap<double[][], double[]> cache = new HashMap<>();
+    private final HashMap<float[][], float[]> cache = new HashMap<>();
 
     private final int[] mask;
     private final int countOutputs;
@@ -17,32 +17,32 @@ public class MaxAbsScaler {
         this.sequenceLength = sequenceLength;
     }
 
-    public void fit(ArrayList<double[][]> data, int countInput) {
+    public void fit(ArrayList<float[][]> data, int countInput) {
         if (data == null || data.isEmpty()) throw new IllegalArgumentException("Input list cannot be empty");
         if (data.getFirst().length == 0 || data.getFirst()[0].length == 0)
             throw new IllegalArgumentException("Input array cannot be empty");
 
-        for (double[][] dataArr : data) {
+        for (float[][] dataArr : data) {
             fit(dataArr, countInput);
         }
     }
 
-    public void fit(double[][] input, int countInput) {
+    public void fit(float[][] input, int countInput) {
         if (input.length == 0 || input[0].length == 0)
             throw new IllegalArgumentException("Input array cannot be empty");
         if (sequenceLength != input[0].length)
             throw new IllegalArgumentException("Input array length are not equals");
 
-        double[] max = new double[sequenceLength];
+        float[] max = new float[sequenceLength];
 
         for (int i = 0; i < sequenceLength; i++) {
-            double[] internalArray = new double[countInput];
+            float[] internalArray = new float[countInput];
 
             for (int j = 0; j < internalArray.length; j++) {
                 internalArray[j] = input[j][i];
             }
 
-            double maxValue = maxAbsValue(internalArray);
+            float maxValue = maxAbsValue(internalArray);
 
             if (Double.isNaN(maxValue) || Double.isInfinite(maxValue)) {
                 throw new IllegalStateException();
@@ -54,16 +54,16 @@ public class MaxAbsScaler {
         this.cache.put(input, max);
     }
 
-    public void transform(ArrayList<double[][]> inputList) {
+    public void transform(ArrayList<float[][]> inputList) {
         if (inputList == null || inputList.isEmpty() || inputList.getFirst().length == 0 || inputList.getFirst()[0].length == 0)
             throw new IllegalArgumentException("Input list cannot be empty");
 
-        for (double[][] input : inputList) {
+        for (float[][] input : inputList) {
             transform(input);
         }
     }
 
-    public void transform(double[][] input) {
+    public void transform(float[][] input) {
         if (input == null || input.length == 0 || input[0].length == 0)
             throw new IllegalArgumentException("Input array cannot be empty");
         if (!cache.containsKey(input))
@@ -71,11 +71,11 @@ public class MaxAbsScaler {
         if (sequenceLength != input[0].length)
             throw new IllegalArgumentException("Input array length are not equals");
 
-        double[] maxAbsValues = this.cache.get(input);
+        float[] maxAbsValues = this.cache.get(input);
 
         for (int i = 0; i < sequenceLength; i++) {
             for (int j = 0; j < input.length; j++) {
-                double value = input[j][i] / maxAbsValues[i];
+                float value = input[j][i] / maxAbsValues[i];
 
                 if (Double.isNaN(value) || Double.isInfinite(value)) {
                     throw new ArithmeticException("Value is NaN");
@@ -86,7 +86,7 @@ public class MaxAbsScaler {
         }
     }
 
-    public void revertFeatures(double[][] input) {
+    public void revertFeatures(float[][] input) {
         if (input == null || input.length == 0 || input[0].length == 0)
             throw new IllegalArgumentException("Input array cannot be empty");
         if (!cache.containsKey(input))
@@ -94,7 +94,7 @@ public class MaxAbsScaler {
         if (sequenceLength != input[0].length)
             throw new IllegalArgumentException("Input array length are not equals");
 
-        double[] maxAbsValues = this.cache.get(input);
+        float[] maxAbsValues = this.cache.get(input);
 
         for (int i = 0; i < sequenceLength; i++) {
             for (int j = 0; j < input.length; j++) {
@@ -103,7 +103,7 @@ public class MaxAbsScaler {
         }
     }
 
-    public void revertLabels(double[][] key, double[][] input) {
+    public void revertLabels(float[][] key, float[][] input) {
         if (input == null || input.length == 0 || input[0].length == 0)
             throw new IllegalArgumentException("Input array cannot be empty");
         if (key == null || key.length == 0)
@@ -111,7 +111,7 @@ public class MaxAbsScaler {
         if (!cache.containsKey(key))
             throw new IllegalArgumentException("Normalizer doesnt has stats for this array");
 
-        double[] maxAbsValues = this.cache.get(key);
+        float[] maxAbsValues = this.cache.get(key);
 
         for (int i = 0; i < mask.length; i++) {
             for (int j = 0; j < countOutputs; j++) {
@@ -120,20 +120,20 @@ public class MaxAbsScaler {
         }
     }
 
-    public void changeBinding(double[][] original, double[][] newBinding) {
+    public void changeBinding(float[][] original, float[][] newBinding) {
         if (!this.cache.containsKey(original))
             throw new IllegalArgumentException("Normalizer doesnt has stats for this array");
 
-        double[] max = this.cache.get(original);
+        float[] max = this.cache.get(original);
 
         this.cache.remove(original);
 
         this.cache.put(newBinding, max);
     }
 
-    public double maxAbsValue(double[] data) {
-        double maxAbsValue = 0;
-        for (double num : data) {
+    public float maxAbsValue(float[] data) {
+        float maxAbsValue = 0;
+        for (float num : data) {
             maxAbsValue = Math.max(maxAbsValue, Math.abs(num));
         }
         return maxAbsValue;
