+++ src/main/java/com/crypto/analysis/main/core/data_utils/normalizers/Transposer.java
@@ -1,8 +1,8 @@
 package com.crypto.analysis.main.core.data_utils.normalizers;
 
 public class Transposer {
-    public static double[][] transpose(double[][] input) {
-        double[][] output = new double[input[0].length][input.length];
+    public static float[][] transpose(float[][] input) {
+        float[][] output = new float[input[0].length][input.length];
         for (int i = 0; i < input.length; i++) {
             for (int j = 0; j < input[0].length; j++) {
                 output[j][i] = input[i][j];
@@ -11,8 +11,8 @@ public class Transposer {
         return output;
     }
 
-    public static double[][] transpose(double[][] input, int length) {
-        double[][] output = new double[input[0].length][length];
+    public static float[][] transpose(float[][] input, int length) {
+        float[][] output = new float[input[0].length][length];
         for (int i = 0; i < input.length; i++) {
             for (int j = 0; j < input[0].length; j++) {
                 output[j][i] = input[i][j];
@@ -21,8 +21,8 @@ public class Transposer {
         return output;
     }
 
-    public static double[][] transposeSingle(double[] input, int length) {
-        double[][] output = new double[input.length][length];
+    public static float[][] transposeSingle(float[] input, int length) {
+        float[][] output = new float[input.length][length];
         for (int i = 0; i < input.length; i++) {
             output[i][0] = input[i];
         }
